+++ src/plugins/metadataSelector/no/noMetadataSelector.interface.ts
@@ -1,5 +1,4 @@
-import {GridMetadata} from '../../../components/metadata';
-import {MetadataSelector, MetadataSelectorOptions} from '../metadataSelector.interface';
+import {GridMetadata, MetadataSelector, MetadataSelectorOptions} from '../../../interfaces';
 
 /**
  * No metadata selector options
@@ -11,6 +10,6 @@ export interface NoMetadataSelectorOptions extends MetadataSelectorOptions
 /**
  * Public API for 'NoMetadataSelectorComponent'
  */
-export interface NoMetadataSelector<TMetadata extends GridMetadata = any> extends MetadataSelector<TMetadata>
+export interface NoMetadataSelector<TMetadata extends GridMetadata = GridMetadata> extends MetadataSelector<TMetadata>
 {
 }